+++ src/Storages/ObjectStorage/DataLakes/Iceberg/IcebergWrites.cpp
@@ -35,6 +35,7 @@
 #include <Common/quoteString.h>
 #include <Common/randomSeed.h>
 #include <Storages/ObjectStorage/DataLakes/Iceberg/ManifestFile.h>
+#include <Storages/ObjectStorage/DataLakes/Iceberg/MetadataGenerator.h>
 #include <Columns/IColumn.h>
 #include <boost/algorithm/string/case_conv.hpp>
 #include <sys/stat.h>
@@ -190,182 +191,6 @@ bool canWriteStatistics(
     return true;
 }
 
-Poco::JSON::Object::Ptr deepCopy(Poco::JSON::Object::Ptr obj)
-{
-    std::ostringstream oss; // STYLE_CHECK_ALLOW_STD_STRING_STREAM
-    obj->stringify(oss);
-
-    Poco::JSON::Parser parser;
-    auto result = parser.parse(oss.str());
-    return result.extract<Poco::JSON::Object::Ptr>();
-}
-
-bool checkValidSchemaEvolution(Poco::Dynamic::Var old_type, Poco::Dynamic::Var new_type)
-{
-    if (old_type.isString() && new_type.isString() && old_type.extract<String>() == new_type.extract<String>())
-        return true;
-
-    if (new_type.isString() && new_type.extract<String>() == "long" &&
-        old_type.isString() && (old_type.extract<String>() == "long" ||  old_type.extract<String>() == "int"))
-    {
-        return true;
-    }
-
-    if (new_type.isString() && new_type.extract<String>() == "double" &&
-        old_type.isString() && (old_type.extract<String>() == "float" ||  old_type.extract<String>() == "double"))
-    {
-        return true;
-    }
-
-    {
-        auto old_complex_type = old_type.extract<Poco::JSON::Object::Ptr>();
-        auto new_complex_type = new_type.extract<Poco::JSON::Object::Ptr>();
-
-        if (old_complex_type && new_complex_type && old_complex_type->has("precision") && new_complex_type->has("precision") &&
-            (old_complex_type->getValue<Int32>("precision") <= new_complex_type->getValue<Int32>("precision") &&
-             old_complex_type->getValue<Int32>("scale") <= new_complex_type->getValue<Int32>("scale")))
-        {
-            return true;
-        }
-    }
-
-    return false;
-}
-
-}
-
-FileNamesGenerator::FileNamesGenerator(
-    const String & table_dir_,
-    const String & storage_dir_,
-    bool use_uuid_in_metadata_,
-    CompressionMethod compression_method_,
-    const String & format_name_)
-    : table_dir(table_dir_)
-    , storage_dir(storage_dir_)
-    , data_dir(table_dir + "data/")
-    , metadata_dir(table_dir + "metadata/")
-    , storage_data_dir(storage_dir + "data/")
-    , storage_metadata_dir(storage_dir + "metadata/")
-    , use_uuid_in_metadata(use_uuid_in_metadata_)
-    , compression_method(compression_method_)
-    , format_name(boost::to_lower_copy(format_name_))
-{
-}
-
-FileNamesGenerator::FileNamesGenerator(const FileNamesGenerator & other)
-{
-    data_dir = other.data_dir;
-    metadata_dir = other.metadata_dir;
-    storage_data_dir = other.storage_data_dir;
-    storage_metadata_dir = other.storage_metadata_dir;
-    initial_version = other.initial_version;
-
-    table_dir = other.table_dir;
-    storage_dir = other.storage_dir;
-    use_uuid_in_metadata = other.use_uuid_in_metadata;
-    compression_method = other.compression_method;
-    format_name = other.format_name;
-}
-
-FileNamesGenerator & FileNamesGenerator::operator=(const FileNamesGenerator & other)
-{
-    if (this == &other)
-        return *this;
-
-    data_dir = other.data_dir;
-    metadata_dir = other.metadata_dir;
-    storage_data_dir = other.storage_data_dir;
-    storage_metadata_dir = other.storage_metadata_dir;
-    initial_version = other.initial_version;
-
-    table_dir = other.table_dir;
-    storage_dir = other.storage_dir;
-    use_uuid_in_metadata = other.use_uuid_in_metadata;
-    compression_method = other.compression_method;
-    format_name = other.format_name;
-
-    return *this;
-}
-
-FileNamesGenerator::Result FileNamesGenerator::generateDataFileName()
-{
-    auto uuid_str = uuid_generator.createRandom().toString();
-
-    return Result{
-        .path_in_metadata = fmt::format("{}data-{}.{}", data_dir, uuid_str, format_name),
-        .path_in_storage = fmt::format("{}data-{}.{}", storage_data_dir, uuid_str, format_name)
-    };
-}
-
-FileNamesGenerator::Result FileNamesGenerator::generateManifestEntryName()
-{
-    auto uuid_str = uuid_generator.createRandom().toString();
-
-    return Result{
-        .path_in_metadata = fmt::format("{}{}.avro", metadata_dir, uuid_str),
-        .path_in_storage = fmt::format("{}{}.avro", storage_metadata_dir, uuid_str),
-    };
-}
-
-FileNamesGenerator::Result FileNamesGenerator::generateManifestListName(Int64 snapshot_id, Int32 format_version)
-{
-    auto uuid_str = uuid_generator.createRandom().toString();
-
-    return Result{
-        .path_in_metadata = fmt::format("{}snap-{}-{}-{}.avro", metadata_dir, snapshot_id, format_version, uuid_str),
-        .path_in_storage = fmt::format("{}snap-{}-{}-{}.avro", storage_metadata_dir, snapshot_id, format_version, uuid_str),
-    };
-}
-
-FileNamesGenerator::Result FileNamesGenerator::generateMetadataName()
-{
-    auto compression_suffix = toContentEncodingName(compression_method);
-    if (!compression_suffix.empty())
-        compression_suffix = "." + compression_suffix;
-    if (!use_uuid_in_metadata)
-    {
-        auto res = Result{
-            .path_in_metadata = fmt::format("{}v{}{}.metadata.json", metadata_dir, initial_version, compression_suffix),
-            .path_in_storage = fmt::format("{}v{}{}.metadata.json", storage_metadata_dir, initial_version, compression_suffix),
-        };
-        initial_version++;
-        return res;
-    }
-    else
-    {
-        auto uuid_str = uuid_generator.createRandom().toString();
-        auto res = Result{
-            .path_in_metadata = fmt::format("{}v{}-{}{}.metadata.json", metadata_dir, initial_version, uuid_str, compression_suffix),
-            .path_in_storage = fmt::format("{}v{}-{}{}.metadata.json", storage_metadata_dir, initial_version, uuid_str, compression_suffix),
-        };
-        initial_version++;
-        return res;
-    }
-}
-
-FileNamesGenerator::Result FileNamesGenerator::generateVersionHint()
-{
-    return Result{
-        .path_in_metadata = fmt::format("{}version-hint.text", metadata_dir),
-        .path_in_storage = fmt::format("{}version-hint.text", storage_metadata_dir),
-    };
-}
-
-FileNamesGenerator::Result FileNamesGenerator::generatePositionDeleteFile()
-{
-    auto uuid_str = uuid_generator.createRandom().toString();
-
-    return Result{
-        .path_in_metadata = fmt::format("{}{}-deletes.{}", data_dir, uuid_str, format_name),
-        .path_in_storage = fmt::format("{}{}-deletes.{}", storage_data_dir, uuid_str, format_name)
-    };
-}
-
-String FileNamesGenerator::convertMetadataPathToStoragePath(const String & metadata_path) const
-{
-    if (!metadata_path.starts_with(table_dir))
-        throw Exception(ErrorCodes::BAD_ARGUMENTS, "Paths in Iceberg must use a consistent format — either /your/path or s3://your/path. Use the write_full_path_in_iceberg_metadata setting to control this behavior {} {}", metadata_path, table_dir);
-    return storage_dir + metadata_path.substr(table_dir.size());
 }
 
 String removeEscapedSlashes(const String & json_str)
@@ -732,584 +557,6 @@ void generateManifestList(
     }
 
     writer.close();
-}
-
-MetadataGenerator::MetadataGenerator(Poco::JSON::Object::Ptr metadata_object_)
-    : metadata_object(metadata_object_)
-    , gen(randomSeed())
-    , dis(0, INT32_MAX)
-{
-}
-
-Int64 MetadataGenerator::getMaxSequenceNumber()
-{
-    auto snapshots = metadata_object->get(Iceberg::f_snapshots).extract<Poco::JSON::Array::Ptr>();
-    Int64 max_seq_number = 0;
-
-    for (size_t i = 0; i < snapshots->size(); ++i)
-    {
-        const auto snapshot = snapshots->getObject(static_cast<UInt32>(i));
-        auto seq_number = snapshot->getValue<Int64>(Iceberg::f_metadata_sequence_number);
-        max_seq_number = std::max(max_seq_number, seq_number);
-    }
-    return max_seq_number;
-}
-
-Poco::JSON::Object::Ptr MetadataGenerator::getParentSnapshot(Int64 parent_snapshot_id)
-{
-    auto snapshots = metadata_object->get(Iceberg::f_snapshots).extract<Poco::JSON::Array::Ptr>();
-    for (size_t i = 0; i < snapshots->size(); ++i)
-    {
-        const auto snapshot = snapshots->getObject(static_cast<UInt32>(i));
-        auto snapshot_id = snapshot->getValue<Int64>(Iceberg::f_metadata_snapshot_id);
-        if (snapshot_id == parent_snapshot_id)
-            return snapshot;
-    }
-    return nullptr;
-}
-
-MetadataGenerator::NextMetadataResult MetadataGenerator::generateNextMetadata(
-    FileNamesGenerator & generator,
-    const String & metadata_filename,
-    Int64 parent_snapshot_id,
-    Int32 added_files,
-    Int32 added_records,
-    Int32 added_files_size,
-    Int32 num_partitions,
-    Int32 added_delete_files,
-    Int32 num_deleted_rows,
-    std::optional<Int64> user_defined_snapshot_id,
-    std::optional<Int64> user_defined_timestamp)
-{
-    int format_version = metadata_object->getValue<Int32>(Iceberg::f_format_version);
-    Poco::JSON::Object::Ptr new_snapshot = new Poco::JSON::Object;
-    if (format_version > 1)
-    {
-        auto sequence_number = getMaxSequenceNumber() + 1;
-        new_snapshot->set(Iceberg::f_metadata_sequence_number, getMaxSequenceNumber() + 1);
-        metadata_object->set(Iceberg::f_last_sequence_number, sequence_number);
-    }
-    Int64 snapshot_id = user_defined_snapshot_id.value_or(static_cast<Int64>(dis(gen)));
-
-    auto [manifest_list_name, storage_manifest_list_name] = generator.generateManifestListName(snapshot_id, format_version);
-    new_snapshot->set(Iceberg::f_metadata_snapshot_id, snapshot_id);
-    new_snapshot->set(Iceberg::f_parent_snapshot_id, parent_snapshot_id);
-
-    auto now = std::chrono::system_clock::now();
-    auto ms = duration_cast<std::chrono::milliseconds>(now.time_since_epoch());
-    Int64 timestamp = user_defined_timestamp.value_or(ms.count());
-    new_snapshot->set(Iceberg::f_timestamp_ms, timestamp);
-    metadata_object->set(Iceberg::f_last_updated_ms, timestamp);
-
-    auto parent_snapshot = getParentSnapshot(parent_snapshot_id);
-    Poco::JSON::Object::Ptr summary = new Poco::JSON::Object;
-    if (num_deleted_rows == 0)
-    {
-        summary->set(Iceberg::f_operation, Iceberg::f_append);
-        summary->set(Iceberg::f_added_data_files, std::to_string(added_files));
-        summary->set(Iceberg::f_added_records, std::to_string(added_records));
-        summary->set(Iceberg::f_added_files_size, std::to_string(added_files_size));
-        summary->set(Iceberg::f_changed_partition_count, std::to_string(num_partitions));
-    }
-    else
-    {
-        summary->set(Iceberg::f_operation, Iceberg::f_overwrite);
-        summary->set(Iceberg::f_added_delete_files, std::to_string(added_delete_files));
-        summary->set(Iceberg::f_added_position_delete_files, std::to_string(added_delete_files));
-        summary->set(Iceberg::f_added_files_size, std::to_string(added_files_size));
-        summary->set(Iceberg::f_added_position_deletes, std::to_string(num_deleted_rows));
-        summary->set(Iceberg::f_changed_partition_count, std::to_string(num_partitions));
-    }
-
-    auto sum_with_parent_snapshot = [&](const char * field_name, Int32 snapshot_value)
-    {
-        Int32 prev_value = parent_snapshot ? std::stoi(parent_snapshot->getObject(Iceberg::f_summary)->getValue<String>(field_name)) : 0;
-        summary->set(field_name, std::to_string(prev_value + snapshot_value));
-    };
-
-    sum_with_parent_snapshot(Iceberg::f_total_records, added_records);
-    sum_with_parent_snapshot(Iceberg::f_total_files_size, added_files_size);
-    sum_with_parent_snapshot(Iceberg::f_total_data_files, added_files);
-    sum_with_parent_snapshot(Iceberg::f_total_delete_files, added_delete_files);
-    sum_with_parent_snapshot(Iceberg::f_total_position_deletes, num_deleted_rows);
-    sum_with_parent_snapshot(Iceberg::f_total_equality_deletes, 0);
-    new_snapshot->set(Iceberg::f_summary, summary);
-
-    new_snapshot->set(Iceberg::f_schema_id, metadata_object->getValue<Int32>(Iceberg::f_current_schema_id));
-    new_snapshot->set(Iceberg::f_manifest_list, manifest_list_name);
-
-    metadata_object->getArray(Iceberg::f_snapshots)->add(new_snapshot);
-    metadata_object->set(Iceberg::f_current_snapshot_id, snapshot_id);
-
-    if (!metadata_object->has(Iceberg::f_refs))
-        metadata_object->set(Iceberg::f_refs, new Poco::JSON::Object);
-
-    if (!metadata_object->getObject(Iceberg::f_refs)->has(Iceberg::f_main))
-    {
-        Poco::JSON::Object::Ptr branch = new Poco::JSON::Object;
-        branch->set(Iceberg::f_metadata_snapshot_id, snapshot_id);
-        branch->set(Iceberg::f_type, Iceberg::f_branch);
-
-        metadata_object->getObject(Iceberg::f_refs)->set(Iceberg::f_main, branch);
-    }
-    else
-        metadata_object->getObject(Iceberg::f_refs)->getObject(Iceberg::f_main)->set(Iceberg::f_metadata_snapshot_id, snapshot_id);
-
-    {
-        Poco::JSON::Object::Ptr new_metadata_item = new Poco::JSON::Object;
-        new_metadata_item->set(Iceberg::f_metadata_file, metadata_filename);
-        new_metadata_item->set(Iceberg::f_timestamp_ms, timestamp);
-        metadata_object->getArray(Iceberg::f_metadata_log)->add(new_metadata_item);
-    }
-    {
-        Poco::JSON::Object::Ptr new_snapshot_item = new Poco::JSON::Object;
-        new_snapshot_item->set(Iceberg::f_metadata_snapshot_id, snapshot_id);
-        new_snapshot_item->set(Iceberg::f_timestamp_ms, timestamp);
-        metadata_object->getArray(Iceberg::f_snapshot_log)->add(new_snapshot_item);
-    }
-
-    if (added_delete_files > 0)
-    {
-        if (!metadata_object->has(Iceberg::f_properties))
-        {
-            Poco::JSON::Object::Ptr properties = new Poco::JSON::Object;
-            metadata_object->set(Iceberg::f_properties, properties);
-        }
-        auto properties = metadata_object->getObject(Iceberg::f_properties);
-        properties->set("owner", "root");
-        properties->set("write.delete.mode", "merge-on-read");
-        properties->set("write.merge.mode", "merge-on-read");
-        properties->set("write.update.mode", "merge-on-read");
-    }
-    return {new_snapshot, manifest_list_name, storage_manifest_list_name};
-}
-
-void MetadataGenerator::generateDropColumnMetadata(const String & column_name)
-{
-    auto current_schema_id = metadata_object->getValue<Int32>(Iceberg::f_current_schema_id);
-    metadata_object->set(Iceberg::f_current_schema_id, current_schema_id + 1);
-
-    Poco::JSON::Object::Ptr current_schema;
-    auto schemas = metadata_object->getArray(Iceberg::f_schemas);
-    for (UInt32 i = 0; i < schemas->size(); ++i)
-    {
-        if (schemas->getObject(i)->getValue<Int32>(Iceberg::f_schema_id) == current_schema_id)
-        {
-            current_schema = schemas->getObject(i);
-            break;
-        }
-    }
-
-    if (!current_schema)
-        throw Exception(ErrorCodes::BAD_ARGUMENTS, "Not found schema with id {}", current_schema_id);
-    current_schema = deepCopy(current_schema);
-
-    auto fields = current_schema->getArray(Iceberg::f_fields);
-    UInt32 index_to_drop = static_cast<UInt32>(fields->size());
-    for (UInt32 i = 0; i < fields->size(); ++i)
-    {
-        if (fields->getObject(i)->getValue<String>(Iceberg::f_name) == column_name)
-        {
-            index_to_drop = i;
-            break;
-        }
-    }
-    if (index_to_drop == fields->size())
-        throw Exception(ErrorCodes::BAD_ARGUMENTS, "Not found column {}", column_name);
-    current_schema->getArray(Iceberg::f_fields)->remove(index_to_drop);
-    current_schema->set(Iceberg::f_schema_id, current_schema_id + 1);
-    metadata_object->getArray(Iceberg::f_schemas)->add(current_schema);
-}
-
-void MetadataGenerator::generateAddColumnMetadata(const String & column_name, DataTypePtr type)
-{
-    if (!type->isNullable())
-        throw Exception(ErrorCodes::BAD_ARGUMENTS, "Iceberg spec doesn't allow to add non-nullable columns");
-    auto current_schema_id = metadata_object->getValue<Int32>(Iceberg::f_current_schema_id);
-    metadata_object->set(Iceberg::f_current_schema_id, current_schema_id + 1);
-
-    Poco::JSON::Object::Ptr current_schema;
-    auto schemas = metadata_object->getArray(Iceberg::f_schemas);
-    for (UInt32 i = 0; i < schemas->size(); ++i)
-    {
-        if (schemas->getObject(i)->getValue<Int32>(Iceberg::f_schema_id) == current_schema_id)
-        {
-            current_schema = schemas->getObject(i);
-            break;
-        }
-    }
-
-    if (!current_schema)
-        throw Exception(ErrorCodes::BAD_ARGUMENTS, "Not found schema with id {}", current_schema_id);
-    current_schema = deepCopy(current_schema);
-    auto last_column_id = metadata_object->getValue<Int32>(Iceberg::f_last_column_id);
-    metadata_object->set(Iceberg::f_last_column_id, last_column_id + 1);
-
-    auto new_type = getIcebergType(type, last_column_id);
-    Poco::JSON::Object::Ptr new_field = new Poco::JSON::Object;
-    new_field->set(Iceberg::f_id, last_column_id + 1);
-    new_field->set(Iceberg::f_name, column_name);
-    new_field->set(Iceberg::f_required, new_type.second);
-    new_field->set(Iceberg::f_type, new_type.first);
-
-    current_schema->getArray(Iceberg::f_fields)->add(new_field);
-    current_schema->set(Iceberg::f_schema_id, current_schema_id + 1);
-    metadata_object->getArray(Iceberg::f_schemas)->add(current_schema);
-}
-
-void MetadataGenerator::generateModifyColumnMetadata(const String & column_name, DataTypePtr type)
-{
-    auto current_schema_id = metadata_object->getValue<Int32>(Iceberg::f_current_schema_id);
-    metadata_object->set(Iceberg::f_current_schema_id, current_schema_id + 1);
-
-    Poco::JSON::Object::Ptr current_schema;
-    auto schemas = metadata_object->getArray(Iceberg::f_schemas);
-    for (UInt32 i = 0; i < schemas->size(); ++i)
-    {
-        if (schemas->getObject(i)->getValue<Int32>(Iceberg::f_schema_id) == current_schema_id)
-        {
-            current_schema = schemas->getObject(i);
-            break;
-        }
-    }
-
-    if (!current_schema)
-        throw Exception(ErrorCodes::BAD_ARGUMENTS, "Not found schema with id {}", current_schema_id);
-    current_schema = deepCopy(current_schema);
-    auto last_column_id = metadata_object->getValue<Int32>(Iceberg::f_last_column_id);
-
-    auto new_type = getIcebergType(type, last_column_id);
-    auto schema_fields = current_schema->getArray(Iceberg::f_fields);
-
-    for (UInt32 i = 0; i < schema_fields->size(); ++i)
-    {
-        auto current_field = schema_fields->getObject(i);
-        if (current_field->getValue<String>(Iceberg::f_name) == column_name)
-        {
-            if (!checkValidSchemaEvolution(current_field->get(Iceberg::f_type), new_type.first))
-                throw Exception(ErrorCodes::BAD_ARGUMENTS, "Iceberg spec doesn't allow schema evolution to type {}", type->getPrettyName());
-
-            auto old_type = deepCopy(current_field);
-            current_field->set(Iceberg::f_type, new_type.first);
-            if (!current_field->getValue<bool>(Iceberg::f_required) && !type->isNullable())
-                throw Exception(ErrorCodes::BAD_ARGUMENTS, "Iceberg spec doesn't allow change type from nullable to non-nullable {}", type->getPrettyName());
-
-            current_field->set(Iceberg::f_required, new_type.second);
-            break;
-        }
-    }
-    current_schema->set(Iceberg::f_schema_id, current_schema_id + 1);
-    metadata_object->getArray(Iceberg::f_schemas)->add(current_schema);
-}
-
-ChunkPartitioner::ChunkPartitioner(
-    Poco::JSON::Array::Ptr partition_specification, Poco::JSON::Object::Ptr schema, ContextPtr context, SharedHeader sample_block_)
-    : sample_block(sample_block_)
-{
-    std::unordered_map<Int32, String> id_to_column;
-    {
-        auto schema_fields = schema->getArray(Iceberg::f_fields);
-        for (size_t i = 0; i < schema_fields->size(); ++i)
-        {
-            auto field = schema_fields->getObject(static_cast<UInt32>(i));
-            id_to_column[field->getValue<Int32>(Iceberg::f_id)] = field->getValue<String>(Iceberg::f_name);
-        }
-    }
-    for (size_t i = 0; i != partition_specification->size(); ++i)
-    {
-        auto partition_specification_field = partition_specification->getObject(static_cast<UInt32>(i));
-
-        auto transform_name = partition_specification_field->getValue<String>("transform");
-        transform_name = Poco::toLower(transform_name);
-
-        FunctionOverloadResolverPtr transform;
-
-        auto source_id = partition_specification_field->getValue<Int32>(Iceberg::f_source_id);
-        auto column_name = id_to_column[source_id];
-
-        auto & factory = FunctionFactory::instance();
-
-        auto transform_and_argument = Iceberg::parseTransformAndArgument(transform_name);
-        if (!transform_and_argument)
-            throw Exception(ErrorCodes::BAD_ARGUMENTS, "Unknown transform {}", transform_name);
-
-        auto function = factory.get(transform_and_argument->transform_name, context);
-
-        ColumnsWithTypeAndName columns_for_function;
-        if (transform_and_argument->argument)
-            columns_for_function.push_back(ColumnWithTypeAndName(nullptr, std::make_shared<DataTypeUInt64>(), ""));
-        columns_for_function.push_back(sample_block_->getByName(column_name));
-
-        result_data_types.push_back(function->getReturnType(columns_for_function));
-        functions.push_back(function);
-        function_params.push_back(transform_and_argument->argument);
-        columns_to_apply.push_back(column_name);
-    }
-}
-
-size_t ChunkPartitioner::PartitionKeyHasher::operator()(const PartitionKey & key) const
-{
-    size_t result = 0;
-    for (const auto & part_key : key)
-        result ^= hasher(part_key.dump());
-    return result;
-}
-
-std::vector<std::pair<ChunkPartitioner::PartitionKey, Chunk>>
-ChunkPartitioner::partitionChunk(const Chunk & chunk)
-{
-    std::unordered_map<String, ColumnWithTypeAndName> name_to_column;
-    for (size_t i = 0; i < sample_block->columns(); ++i)
-    {
-        auto column_ptr = chunk.getColumns()[i];
-        auto column_name = sample_block->getNames()[i];
-        name_to_column[column_name] = ColumnWithTypeAndName(column_ptr, sample_block->getDataTypes()[i], column_name);
-    }
-
-    std::vector<ChunkPartitioner::PartitionKey> transform_results(chunk.getNumRows());
-    for (size_t transform_ind = 0; transform_ind < functions.size(); ++transform_ind)
-    {
-        ColumnsWithTypeAndName arguments;
-        if (function_params[transform_ind].has_value())
-        {
-            auto type = std::make_shared<DataTypeUInt64>();
-            auto column_value = ColumnUInt64::create();
-            column_value->insert(*function_params[transform_ind]);
-            auto const_column = ColumnConst::create(std::move(column_value), chunk.getNumRows());
-            arguments.push_back(ColumnWithTypeAndName(const_column->clone(), type, "#"));
-        }
-        arguments.push_back(name_to_column[columns_to_apply[transform_ind]]);
-        auto result
-            = functions[transform_ind]->build(arguments)->execute(arguments, std::make_shared<DataTypeString>(), chunk.getNumRows(), false);
-        for (size_t i = 0; i < chunk.getNumRows(); ++i)
-        {
-            Field field;
-            result->get(i, field);
-            transform_results[i].push_back(field);
-        }
-    }
-
-    auto get_partition = [&](size_t row_num)
-    {
-        return transform_results[row_num];
-    };
-
-    PODArray<size_t> partition_num_to_first_row;
-    IColumn::Selector selector;
-    ColumnRawPtrs raw_columns;
-    for (const auto & column : chunk.getColumns())
-        raw_columns.push_back(column.get());
-
-    buildScatterSelector(raw_columns, partition_num_to_first_row, selector, 0, Context::getGlobalContextInstance());
-
-
-    size_t partitions_count = partition_num_to_first_row.size();
-    chassert(partitions_count > 0);
-    std::vector<std::pair<ChunkPartitioner::PartitionKey, MutableColumns>> result_columns;
-    result_columns.reserve(partitions_count);
-
-    for (size_t i = 0; i < partitions_count; ++i)
-        result_columns.push_back({get_partition(partition_num_to_first_row[i]), chunk.cloneEmptyColumns()});
-
-    for (size_t col = 0; col < chunk.getNumColumns(); ++col)
-    {
-        if (partitions_count > 1)
-        {
-            MutableColumns scattered = chunk.getColumns()[col]->scatter(partitions_count, selector);
-            for (size_t i = 0; i < partitions_count; ++i)
-                result_columns[i].second[col] = std::move(scattered[i]);
-        }
-        else
-        {
-            result_columns[0].second[col] = chunk.getColumns()[col]->cloneFinalized();
-        }
-    }
-
-    std::vector<std::pair<ChunkPartitioner::PartitionKey, Chunk>> result;
-    result.reserve(result_columns.size());
-    for (auto && [key, partition_columns] : result_columns)
-    {
-        size_t column_size = partition_columns[0]->size();
-        result.push_back({key, Chunk(std::move(partition_columns), column_size)});
-    }
-    return result;
-}
-
-DataFileStatistics::DataFileStatistics(Poco::JSON::Array::Ptr schema_)
-{
-    field_ids.resize(schema_->size());
-    for (UInt32 i = 0; i < schema_->size(); ++i)
-    {
-        auto field = schema_->getObject(i);
-        size_t field_id = field->getValue<size_t>(Iceberg::f_id);
-        field_ids[i] =  field_id;
-    }
-}
-
-Range getExtremeRangeFromColumn(const ColumnPtr & column)
-{
-    Field min_val;
-    Field max_val;
-    column->getExtremes(min_val, max_val);
-    return Range(min_val, true, max_val, true);
-}
-
-void DataFileStatistics::update(const Chunk & chunk)
-{
-    size_t num_columns = chunk.getNumColumns();
-    if (column_sizes.empty())
-    {
-        column_sizes.resize(num_columns, 0);
-        null_counts.resize(num_columns, 0);
-        for (size_t i = 0; i < num_columns; ++i)
-        {
-            ranges.push_back(getExtremeRangeFromColumn(chunk.getColumns()[i]));
-        }
-    }
-
-    chassert(ranges.size() == num_columns);
-
-    for (size_t i = 0; i < num_columns; ++i)
-    {
-        column_sizes[i] += chunk.getColumns()[i]->byteSize();
-        for (size_t j = 0; j < chunk.getNumRows(); ++j)
-            null_counts[i] += (chunk.getColumns()[i]->isNullAt(j));
-        ranges[i] = uniteRanges(ranges[i], getExtremeRangeFromColumn(chunk.getColumns()[i]));
-    }
-}
-
-Range DataFileStatistics::uniteRanges(const Range & left, const Range & right)
-{
-    return Range(
-        Range::less(left.left, right.left) ? left.left : right.left,
-        true,
-        Range::less(right.right, left.right) ? left.right : right.right,
-        true);
-}
-
-std::vector<std::pair<size_t, size_t>> DataFileStatistics::getColumnSizes() const
-{
-    std::vector<std::pair<size_t, size_t>> result;
-    for (size_t i = 0; i < column_sizes.size(); ++i)
-    {
-        result.push_back({field_ids[i], column_sizes[i]});
-    }
-    return result;
-}
-
-std::vector<std::pair<size_t, size_t>> DataFileStatistics::getNullCounts() const
-{
-    std::vector<std::pair<size_t, size_t>> result;
-    for (size_t i = 0; i < null_counts.size(); ++i)
-    {
-        result.push_back({field_ids[i], null_counts[i]});
-    }
-    return result;
-}
-
-
-std::vector<std::pair<size_t, Field>> DataFileStatistics::getLowerBounds() const
-{
-    std::vector<std::pair<size_t, Field>> result;
-    for (size_t i = 0; i < ranges.size(); ++i)
-    {
-        result.push_back({field_ids[i], ranges[i].left});
-    }
-    return result;
-}
-
-std::vector<std::pair<size_t, Field>> DataFileStatistics::getUpperBounds() const
-{
-    std::vector<std::pair<size_t, Field>> result;
-    for (size_t i = 0; i < ranges.size(); ++i)
-    {
-        result.push_back({field_ids[i], ranges[i].right});
-    }
-    return result;
-}
-
-MultipleFileWriter::MultipleFileWriter(
-    UInt64 max_data_file_num_rows_,
-    UInt64 max_data_file_num_bytes_,
-    Poco::JSON::Array::Ptr schema,
-    FileNamesGenerator & filename_generator_,
-    ObjectStoragePtr object_storage_,
-    ContextPtr context_,
-    const std::optional<FormatSettings> & format_settings_,
-    StorageObjectStorageConfigurationPtr configuration_,
-    SharedHeader sample_block_)
-    : max_data_file_num_rows(max_data_file_num_rows_)
-    , max_data_file_num_bytes(max_data_file_num_bytes_)
-    , stats(schema)
-    , filename_generator(filename_generator_)
-    , object_storage(object_storage_)
-    , context(context_)
-    , format_settings(format_settings_)
-    , configuration(configuration_)
-    , sample_block(sample_block_)
-{
-}
-
-void MultipleFileWriter::consume(const Chunk & chunk)
-{
-    if (!current_file_num_rows || *current_file_num_rows >= max_data_file_num_rows || *current_file_num_bytes >= max_data_file_num_bytes)
-    {
-        if (buffer)
-            finalize();
-
-        current_file_num_rows = 0;
-        current_file_num_bytes = 0;
-        auto filename = filename_generator.generateDataFileName();
-
-        data_file_names.push_back(filename.path_in_storage);
-        buffer = object_storage->writeObject(
-            StoredObject(filename.path_in_storage), WriteMode::Rewrite, std::nullopt, DBMS_DEFAULT_BUFFER_SIZE, context->getWriteSettings());
-
-        if (format_settings)
-        {
-            format_settings->parquet.write_page_index = true;
-            format_settings->parquet.bloom_filter_push_down = true;
-            format_settings->parquet.filter_push_down = true;
-        }
-        output_format = FormatFactory::instance().getOutputFormatParallelIfPossible(
-            configuration->format, *buffer, *sample_block, context, format_settings);
-    }
-    output_format->write(sample_block->cloneWithColumns(chunk.getColumns()));
-    output_format->flush();
-    *current_file_num_rows += chunk.getNumRows();
-    *current_file_num_bytes += chunk.bytes();
-    stats.update(chunk);
-}
-
-void MultipleFileWriter::finalize()
-{
-    output_format->flush();
-    output_format->finalize();
-    buffer->finalize();
-    total_bytes += buffer->count();
-}
-
-void MultipleFileWriter::release()
-{
-    output_format.reset();
-    buffer.reset();
-}
-
-void MultipleFileWriter::cancel()
-{
-    output_format->cancel();
-    buffer->cancel();
-}
-
-void MultipleFileWriter::clearAllDataFiles() const
-{
-    for (const auto & data_filename : data_file_names)
-        object_storage->removeObjectIfExists(StoredObject(data_filename));
-}
-
-UInt64 MultipleFileWriter::getResultBytes() const
-{
-    return total_bytes;
 }
 
 IcebergStorageSink::IcebergStorageSink(
